+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 `tomland` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+## Unrelised
+
+* [#325](https://github.com/kowainik/tomland/issues/325):
+  Add ability to one or multiline printing to `PrintOptions` for arrays.
+
 ## 1.3.1.0 — Sep 21, 2020
 
 * [#331](https://github.com/kowainik/tomland/issues/331):